chore(user-service & restaurant-service): try configuration

+++ restaurant-service/src/main/java/com/serhatacar/restaurantservice/dto/response/RestaurantWithScoreDTO.java
@@ -2,8 +2,6 @@ package com.serhatacar.restaurantservice.dto.response;
 
 import com.serhatacar.restaurantservice.entity.enums.Status;
 
-import java.math.BigDecimal;
-
 /**
  * @author Serhat Acar
  */
@@ -19,7 +17,7 @@ public record RestaurantWithScoreDTO(
         String workingHours,
         double latitude,
         double longitude,
-        BigDecimal restaurantRate,
+        Double restaurantRate,
         double score,
         Status status
